requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Azure Services
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
openai==1.3.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0